--- a/lib/ColorThief/Image/Adapter/GDImageAdapter.php
+++ b/lib/ColorThief/Image/Adapter/GDImageAdapter.php
@@ -72,10 +72,6 @@ class GDImageAdapter extends ImageAdapter
                 $resource = @imagecreatefrompng($file);
                 break;
 
-            case IMAGETYPE_WEBP:
-                $resource = @imagecreatefromwebp($file);
-                break;
-
             default:
                 throw new \RuntimeException("Image '{$file}' is not readable or does not exists.");
                 break;