--- a/lib/ColorThief/VBox.php
+++ b/lib/ColorThief/VBox.php
@@ -46,20 +46,21 @@ class VBox
             // the number of pixels contained in this vbox.
             if ($this->volume() > count($this->histo)) {
                 // Iterate over the histogram if the size of this histogram is lower than the vbox volume
-                foreach ($this->histo as $rgb => $count) {
-                    $rgb_array = ColorThief::getColorsFromIndex($rgb, 0, ColorThief::SIGBITS);
-                    if ($this->contains($rgb_array, 0)) {
+                foreach ($this->histo as $bucketInt => $count) {
+                    $rgbBuckets = ColorThief::getColorsFromIndex($bucketInt, 0, ColorThief::SIGBITS);
+                    if ($this->contains($rgbBuckets, 0)) {
                         $npix += $count;
                     }
                 }
             } else {
                 // Or iterate over points of the vbox if the size of the histogram is greater than the vbox volume
-                for ($i = $this->r1; $i <= $this->r2; $i++) {
-                    for ($j = $this->g1; $j <= $this->g2; $j++) {
-                        for ($k = $this->b1; $k <= $this->b2; $k++) {
-                            $index = ColorThief::getColorIndex($i, $j, $k);
-                            if (isset($this->histo[$index])) {
-                                $npix += $this->histo[$index];
+                for ($redBucket = $this->r1; $redBucket <= $this->r2; $redBucket++) {
+                    for ($greenBucket = $this->g1; $greenBucket <= $this->g2; $greenBucket++) {
+                        for ($blueBucket = $this->b1; $blueBucket <= $this->b2; $blueBucket++) {
+                            // The getColorIndex function takes RGB values instead of buckets. The left shift converts our bucket into its RGB value.
+                            $bucketInt = ColorThief::getColorIndex($redBucket << ColorThief::RSHIFT, $greenBucket << ColorThief::RSHIFT, $blueBucket << ColorThief::RSHIFT, ColorThief::SIGBITS);
+                            if (isset($this->histo[$bucketInt])) {
+                                $npix += $this->histo[$bucketInt];
                             }
                         }
                     }
@@ -77,24 +78,35 @@ class VBox
         return new self($this->r1, $this->r2, $this->g1, $this->g2, $this->b1, $this->b2, $this->histo);
     }
 
+    /**
+     * Calculates the average color represented by this VBox.
+     *
+     * @param bool $force
+     * @return array|bool
+     */
     public function avg($force = false)
     {
         if (!$this->avg || $force) {
             $ntot = 0;
-            $mult = 1 << (8 - ColorThief::SIGBITS);
+            $mult = 1 << ColorThief::RSHIFT;
             $rsum = 0;
             $gsum = 0;
             $bsum = 0;
 
-            for ($i = $this->r1; $i <= $this->r2; $i++) {
-                for ($j = $this->g1; $j <= $this->g2; $j++) {
-                    for ($k = $this->b1; $k <= $this->b2; $k++) {
-                        $histoindex = ColorThief::getColorIndex($i, $j, $k);
-                        $hval = isset($this->histo[$histoindex]) ? $this->histo[$histoindex] : 0;
+            for ($redBucket = $this->r1; $redBucket <= $this->r2; $redBucket++) {
+                for ($greenBucket = $this->g1; $greenBucket <= $this->g2; $greenBucket++) {
+                    for ($blueBucket = $this->b1; $blueBucket <= $this->b2; $blueBucket++) {
+                        // getColorIndex takes RGB values instead of buckets, so left shift so we get a bucketInt
+                        $bucketInt = ColorThief::getColorIndex($redBucket<<3, $greenBucket<<3, $blueBucket<<3, ColorThief::SIGBITS);
+
+                        // The bucket values need to be multiplied by $mult to get the RGB values.
+                        // Normally would use a left shift, but adding 0.5 to put the value at the bucket's midpoint.
+                        // Not really sure what happens if that creates a floating point number, so multiplying is safer.
+                        $hval = isset($this->histo[$bucketInt]) ? $this->histo[$bucketInt] : 0;
                         $ntot += $hval;
-                        $rsum += ($hval * ($i + 0.5) * $mult);
-                        $gsum += ($hval * ($j + 0.5) * $mult);
-                        $bsum += ($hval * ($k + 0.5) * $mult);
+                        $rsum += ($hval * ($redBucket + 0.5) * $mult);
+                        $gsum += ($hval * ($greenBucket + 0.5) * $mult);
+                        $bsum += ($hval * ($blueBucket + 0.5) * $mult);
                     }
                 }
             }
@@ -123,19 +135,20 @@ class VBox
         return $this->avg;
     }
 
-    public function contains(array $pixel, $rshift = ColorThief::RSHIFT)
+    public function contains(array $rgbValue, $rshift = ColorThief::RSHIFT)
     {
-        $rval = $pixel[0] >> $rshift;
-        $gval = $pixel[1] >> $rshift;
-        $bval = $pixel[2] >> $rshift;
+        // Get the buckets from the RGB values.
+        $redBucket = $rgbValue[0] >> $rshift;
+        $greenBucket = $rgbValue[1] >> $rshift;
+        $blueBucket = $rgbValue[2] >> $rshift;
 
         return
-            $rval >= $this->r1 &&
-            $rval <= $this->r2 &&
-            $gval >= $this->g1 &&
-            $gval <= $this->g2 &&
-            $bval >= $this->b1 &&
-            $bval <= $this->b2;
+            $redBucket >= $this->r1 &&
+            $redBucket <= $this->r2 &&
+            $greenBucket >= $this->g1 &&
+            $greenBucket <= $this->g2 &&
+            $blueBucket >= $this->b1 &&
+            $blueBucket <= $this->b2;
     }
 
     /**
@@ -145,10 +158,11 @@ class VBox
      */
     public function longestAxis()
     {
-        $colorWidth['r'] = $this->r2 - $this->r1;
-        $colorWidth['g'] = $this->g2 - $this->g1;
-        $colorWidth['b'] = $this->b2 - $this->b1;
+        // Color-Width for RGB
+        $red = $this->r2 - $this->r1;
+        $green = $this->g2 - $this->g1;
+        $blue = $this->b2 - $this->b1;
 
-        return array_search(max($colorWidth), $colorWidth);
+        return $red >= $green && $red >= $blue ? 'r' : ($green >= $red && $green >= $blue ? 'g' : 'b');
     }
 }