Speeds things up considerably and uses quite a bit less memory when testing the vegetables_1500x995.png test image.

Basic Symfony stopwatch testing says this is faster than the javascript version.  While I hesitate to compare numbers since the hardware and OS is different between my laptop and my dev server, it is much improved.  I think the original timing was around 1000ms.  The javascript version gets about 490ms on their drag-and-drop demo and this is now getting around 110ms when timing `ColorThief::getPalette($thiefImage, 10, 10)` which I think is a fair comparison at comparing a quality of every 10 pixels.   Using a quality of 1 pixel and timing `ColorThief::getPalette($thiefImage, 10, 1)` averages about 520ms.

Memory usage is also drastically improved with stopwatch reporting 2MB and the symfony dev toolbar shows the page using 8MB.  I think this was over 20MB previously.