@mreiden 2018-10-12
This includes or supersedes my other pull requests.

Sorry it's not split up into multiple commits, but I did add comments and renamed variables to try and indicate RGB color values vs histogram buckets, etc.