@ksubileau 2018-10-14
Good job, I've done some tests and indeed I noticed a considerable performance improvement: around 2 times faster and between 20% to 60% of memory saved!  That's really interresting.

But I'm a little confused between your different PRs that seems to overlap partially but not completely. There's also a lot of changes in one chunk with apparently different purpose, so I must admit that it's a little bit hard for me to review.