@mreiden 2018-10-14
The bitwise OR commit was done to be consistent with the loadImage code.

Using a bitwise OR seems to save about 5ms compared to regular addition if you do it often enough (~2.0 million pixels).  Given all the care to make sure bits don't overlap, I thought it might indicate the separation of bits for each color too (RRRRRGGGGGBBBBB).