@mreiden 2018-10-14
I gave up on trying to use separate PRs for things because many of them were dependent on the earlier changes and I noticed there were commits all other the place.

This pull request supersedes #43 and includes #42.   We use git very sparingly and I honestly don't know how to do these individual but dependent pull requests correctly.