@ksubileau 2018-10-27
I finally reworked your PR a bit before merging it, but it's still a very great job with a very clear performance gain. Thank you for your work!